requirements.txt
# 该文件列出了复现视觉发散实验所需的 Python 依赖
# 建议在 Conda 环境内使用 `pip install -r requirements.txt` 安装
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
datasets>=2.15.0
pillow>=10.1.0
peft>=0.7.1
accelerate>=0.24.1
matplotlib>=3.8.0
scikit-learn>=1.3.0
numpy>=1.24.0
pyyaml>=6.0.1
tqdm>=4.66.0

